requirements.txt
torch>=2.0.0
torchvision>=0.15.1
transformers>=4.30.0
ultralytics>=8.0.0
opencv-python>=4.7.0
Pillow>=9.5.0
matplotlib>=3.7.1
numpy>=1.24.3
tabulate>=0.9.0
tqdm>=4.65.0